Day 10
This one was pretty neat. Part 2 was tough, but I wanted to see if I
could get my idea to scale the map to work. I had a surprisingly
difficult time getting the BFS right because there are multiple ways to
get to each tile. In order to debug the program I finally figured out
how to block on input from stdin. The solution I ended up with works and
is really fast which feels good.